+++ cmd/frontend/internal/httpapi/httpapi.go
@@ -182,7 +182,7 @@ func NewHandler(
 	// Return the minimum src-cli version that's compatible with this instance
 	m.Get(apirouter.SrcCli).Handler(trace.Route(newSrcCliVersionHandler(logger)))
 
-	gsClient := gitserver.NewClient(db)
+	gsClient := gitserver.NewClient()
 	m.Get(apirouter.GitBlameStream).Handler(trace.Route(handleStreamBlame(logger, db, gsClient)))
 
 	// Set up the src-cli version cache handler (this will effectively be a
@@ -226,7 +226,7 @@ func RegisterInternalServices(
 	})
 
 	// zoekt-indexserver endpoints
-	gsClient := gitserver.NewClient(db)
+	gsClient := gitserver.NewClient()
 	indexer := &searchIndexerServer{
 		db:              db,
 		logger:          logger.Scoped("searchIndexerServer", "zoekt-indexserver endpoints"),