+++ internal/redispool/ratelimit_test.go
@@ -22,8 +22,8 @@ func TestRateLimiter(t *testing.T) {
 	// Set up the test by initializing the bucket with some initial capacity and replenishment rate
 	ctx := context.Background()
 	bucketName := "github.com:api_tokens"
-	bucketCapacity := 100
-	bucketReplenishRateSeconds := 10
+	bucketCapacity := int32(100)
+	bucketReplenishRateSeconds := int32(10)
 
 	// Try to get tokens before rate limiter config is set in Redis
 	_, _, err := rl.GetTokensFromBucket(ctx, bucketName, 1)
@@ -106,32 +106,9 @@ func redisPoolForTest(t *testing.T, prefix string) *redis.Pool {
 		c.Close()
 	})
 
-	if err := deleteAllKeysWithPrefix(c, prefix); err != nil {
+	if err := DeleteAllKeysWithPrefix(c, prefix); err != nil {
 		t.Logf("Could not clear test prefix name=%q prefix=%q error=%v", t.Name(), prefix, err)
 	}
 
 	return pool
-}
-
-func deleteAllKeysWithPrefix(c redis.Conn, prefix string) error {
-	const script = `
-redis.replicate_commands()
-local cursor = '0'
-local prefix = ARGV[1]
-local batchSize = ARGV[2]
-local result = ''
-repeat
-	local keys = redis.call('SCAN', cursor, 'MATCH', prefix, 'COUNT', batchSize)
-	if #keys[2] > 0
-	then
-		result = redis.call('DEL', unpack(keys[2]))
-	end
-
-	cursor = keys[1]
-until cursor == '0'
-return result
-`
-
-	_, err := c.Do("EVAL", script, 0, prefix+":*", 100)
-	return err
 }