+++ client/jetbrains/CHANGELOG.md
@@ -7,11 +7,13 @@
 - New settings to enable debugging with the agent [#55821](https://github.com/sourcegraph/sourcegraph/pull/55821)
 - Added ability to hide completion suggestions with ESC key [#55955](https://github.com/sourcegraph/sourcegraph/pull/55955)
 - New alt-backslash shortcut to exlicitly trigger autocomplete [#55926](https://github.com/sourcegraph/sourcegraph/pull/55926)
+- Add visual hints about Cody status to status bar [#56046](https://github.com/sourcegraph/sourcegraph/pull/56046)
 
 ### Changed
 
 - Improved settings UI [#55876](https://github.com/sourcegraph/sourcegraph/pull/55876)
 - Telemetry and other GraphQL requests are now sent through the agent [56001](https://github.com/sourcegraph/sourcegraph/pull/56001)
+- Use agent for recipes [#56196](https://github.com/sourcegraph/sourcegraph/pull/56196)
 
 ### Deprecated
 
@@ -23,8 +25,13 @@
 
 ### Fixed
 
-- Await on agent server before submitting telemetry events [56007](https://github.com/sourcegraph/sourcegraph/pull/56007)
-- Bug causing exceptions to get thrown on editor events [55999](https://github.com/sourcegraph/sourcegraph/pull/55999)
+- Handle uncaught exception [#56048](https://github.com/sourcegraph/sourcegraph/pull/56048)
+- Start the agent process on Windows [#56055](https://github.com/sourcegraph/sourcegraph/pull/56055)
+- Internal: use `Autocomplete` instead of `AutoComplete` [#56106](https://github.com/sourcegraph/sourcegraph/pull/56106)
+- Start the agent process on installation events [#56116](https://github.com/sourcegraph/sourcegraph/pull/56116)
+- Cancel outdated autocomplete requests [#56119](https://github.com/sourcegraph/sourcegraph/pull/56119) [sourcegraph/cody#787](https://github.com/sourcegraph/cody/pull/787)
+- Await on agent server before submitting telemetry events [#56007](https://github.com/sourcegraph/sourcegraph/pull/56007)
+- Bug causing exceptions to get thrown on editor events [#55999](https://github.com/sourcegraph/sourcegraph/pull/55999)
 - Use inferred codebase for autocomplete [#55900](https://github.com/sourcegraph/sourcegraph/pull/55900)
 - Make sure caret is visible after accepting multiline completion [#55924](https://github.com/sourcegraph/sourcegraph/pull/55924)
 - Suppress duplicate telemetry when using agent [cody#689](https://github.com/sourcegraph/cody/pull/689)